+++ analyzing_mavlink_message_handler.cpp
@@ -7,8 +7,8 @@ void Analyzing_MAVLink_Message_Handler::end_of_log(uint32_t packet_count)
 
 
 void Analyzing_MAVLink_Message_Handler::handle_decoded_message(uint64_t T, mavlink_ahrs2_t &msg) {
-    _vehicle->position_estimate("AHRS2")->set_lat(T, msg.lat/10000000.0f);
-    _vehicle->position_estimate("AHRS2")->set_lon(T, msg.lng/10000000.0f);
+    _vehicle->position_estimate("AHRS2")->set_lat(T, msg.lat/(double)10000000.0f);
+    _vehicle->position_estimate("AHRS2")->set_lon(T, msg.lng/(double)10000000.0f);
     _vehicle->altitude_estimate("AHRS2")->set_alt(T, msg.altitude);
 
     _vehicle->attitude_estimate("AHRS2")->set_roll(T, rad_to_deg(msg.roll));
@@ -33,27 +33,27 @@ void Analyzing_MAVLink_Message_Handler::handle_decoded_message(uint64_t T, mavli
 }
 
 void Analyzing_MAVLink_Message_Handler::handle_decoded_message(uint64_t T, mavlink_global_position_int_t &msg) {
-    _vehicle->position_estimate("GLOBAL_POSITION_INT")->set_lat(T, msg.lat/10000000.0f);
-    _vehicle->position_estimate("GLOBAL_POSITION_INT")->set_lon(T, msg.lon/10000000.0f);
-    _vehicle->altitude_estimate("GLOBAL_POSITION_INT")->set_alt(T, msg.alt/1000.0f);
+    _vehicle->position_estimate("GLOBAL_POSITION_INT")->set_lat(T, msg.lat/(double)10000000.0f);
+    _vehicle->position_estimate("GLOBAL_POSITION_INT")->set_lon(T, msg.lon/(double)10000000.0f);
+    _vehicle->altitude_estimate("GLOBAL_POSITION_INT")->set_alt(T, msg.alt/(double)1000.0f);
 
     // GLOBAL_POSITION_INT is the "offical" position of the vehicle:
     _vehicle->set_T(T);
-    _vehicle->set_lat(msg.lat/10000000.0f);
-    _vehicle->set_lon(msg.lon/10000000.0f);
+    _vehicle->set_lat(msg.lat/(double)10000000.0f);
+    _vehicle->set_lon(msg.lon/(double)10000000.0f);
     // and I'm assuming for ALT as well:
-    _vehicle->set_altitude(msg.alt/1000.0f);
+    _vehicle->set_altitude(msg.alt/(double)1000.0f);
 
     _analyze->evaluate_all();
 }
 
 void Analyzing_MAVLink_Message_Handler::handle_decoded_message(uint64_t T, mavlink_gps_raw_int_t &msg) {
-    _vehicle->position_estimate("GPS_RAW_INT")->set_lat(T, msg.lat/10000000.0f);
-    _vehicle->position_estimate("GPS_RAW_INT")->set_lon(T, msg.lon/10000000.0f);
-    _vehicle->altitude_estimate("GPS_RAW_INT")->set_alt(T, msg.alt/1000.0f);
+    _vehicle->position_estimate("GPS_RAW_INT")->set_lat(T, msg.lat/(double)10000000.0f);
+    _vehicle->position_estimate("GPS_RAW_INT")->set_lon(T, msg.lon/(double)10000000.0f);
+    _vehicle->altitude_estimate("GPS_RAW_INT")->set_alt(T, msg.alt/(double)1000.0f);
 
     _vehicle->gpsinfo("GPS_RAW_INT")->set_satellites(msg.satellites_visible);
-    _vehicle->gpsinfo("GPS_RAW_INT")->set_hdop(msg.eph/100.0f);
+    _vehicle->gpsinfo("GPS_RAW_INT")->set_hdop(msg.eph/(double)100.0f);
 
     _analyze->evaluate_all();
 }
